Add upstream commit ahead/behind checking for info
This gave the the ability to check if push/pull needs to happen or if we
can forget it (speed boost!). Also, it negates the need for UpToDate
which checks human-readable output, which is dangerous if git is updated
in the future. Therefore, UpToDate function is removed.

This commit brings speed improvements AND more information! :)